requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
python-dotenv==0.7.1
requests==2.18.4
tmdbsimple==2.0.0
urllib3==1.22
